extract token persistence functions to utils
